@@ -1,11 +1,4 @@
-  
-import Vue from 'vue'
+import { createApp } from 'vue'
 import Demo from './Demo.vue'
 
-Vue.config.productionTip = false
-
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  render: h => h(Demo)
-})
+createApp(Demo).mount("#app");
